+++ files/gitlab-cookbooks/gitlab/attributes/default.rb
@@ -254,6 +254,7 @@ default['gitlab']['postgresql']['home'] = "/var/opt/gitlab/postgresql"
 default['gitlab']['postgresql']['user_path'] = "#{node['package']['install-dir']}/embedded/bin:#{node['package']['install-dir']}/bin:$PATH"
 default['gitlab']['postgresql']['sql_user'] = "gitlab"
 default['gitlab']['postgresql']['sql_ci_user'] = "gitlab_ci"
+default['gitlab']['postgresql']['sql_mattermost_user'] = "gitlab_mattermost"
 default['gitlab']['postgresql']['port'] = 5432
 default['gitlab']['postgresql']['listen_address'] = nil
 default['gitlab']['postgresql']['max_connections'] = 200
@@ -515,3 +516,88 @@ default['gitlab']['ci-redis']['unixsocket'] = "/var/opt/gitlab/ci-redis/redis.so
 ####
 default['gitlab']['ci-nginx'] = default['gitlab']['nginx'].dup
 default['gitlab']['ci-nginx']['enable'] = false
+
+####
+# Mattermost
+####
+
+default['gitlab']['mattermost']['enable'] = false
+default['gitlab']['mattermost']['username'] = 'mattermost'
+default['gitlab']['mattermost']['group'] = 'mattermost'
+default['gitlab']['mattermost']['home'] = '/var/opt/gitlab/mattermost'
+default['gitlab']['mattermost']['database_name'] = 'mattermost_production'
+
+default['gitlab']['mattermost']['log_file_directory'] = '/var/log/gitlab/mattermost'
+default['gitlab']['mattermost']['log_console_enable'] = true
+default['gitlab']['mattermost']['log_console_level'] = 'INFO'
+default['gitlab']['mattermost']['log_file_enable'] = false
+default['gitlab']['mattermost']['log_file_level'] = 'INFO'
+default['gitlab']['mattermost']['log_file_format'] = nil
+
+default['gitlab']['mattermost']['service_site_name'] = "GitLab Mattermost"
+default['gitlab']['mattermost']['service_mode'] = 'beta'
+default['gitlab']['mattermost']['service_allow_testing'] = false
+default['gitlab']['mattermost']['service_use_ssl'] = false
+default['gitlab']['mattermost']['service_port'] = "8065"
+default['gitlab']['mattermost']['service_version'] = "developer"
+default['gitlab']['mattermost']['service_analytics_url'] = nil
+default['gitlab']['mattermost']['service_use_local_storage'] = true
+default['gitlab']['mattermost']['service_storage_directory'] = "/var/opt/gitlab/mattermost/data"
+default['gitlab']['mattermost']['service_allowed_login_attempts'] = 10
+
+default['gitlab']['mattermost']['sql_data_source'] = nil
+default['gitlab']['mattermost']['sql_data_source_replicas'] = []
+default['gitlab']['mattermost']['sql_max_idle_conns'] = 10
+default['gitlab']['mattermost']['sql_max_open_conns'] = 10
+default['gitlab']['mattermost']['sql_trace'] = false
+
+# default['gitlab']['mattermost']['oauth'] = {'gitlab' => {'Allow' => true, 'Secret' => "123", 'Id' => "123", "AuthEndpoint" => "aa", "TokenEndpoint" => "bb", "UserApiEndpoint" => "cc" }}
+default['gitlab']['mattermost']['oauth'] = {}
+# default['gitlab']['mattermost']['aws'] = {'S3AccessKeyId' => '123', 'S3SecretAccessKey' => '123', 'S3Bucket' => 'aa', 'S3Region' => 'bb'}
+default['gitlab']['mattermost']['aws'] = {}
+default['gitlab']['mattermost']['image_thumbnail_width'] = 120
+default['gitlab']['mattermost']['image_thumbnail_height'] = 100
+default['gitlab']['mattermost']['image_preview_width'] = 1024
+default['gitlab']['mattermost']['image_preview_height'] = 0
+default['gitlab']['mattermost']['image_profile_width'] = 128
+default['gitlab']['mattermost']['image_profile_height'] = 128
+default['gitlab']['mattermost']['image_initial_font'] = 'luximbi.ttf'
+
+default['gitlab']['mattermost']['email_by_pass_email'] = true
+default['gitlab']['mattermost']['email_smtp_username'] = nil
+default['gitlab']['mattermost']['email_smtp_password'] = nil
+default['gitlab']['mattermost']['email_smtp_server'] = nil
+default['gitlab']['mattermost']['email_use_tls'] = false
+default['gitlab']['mattermost']['email_feedback_email'] = nil
+default['gitlab']['mattermost']['email_feedback_name'] = nil
+default['gitlab']['mattermost']['email_apple_push_server'] = nil
+default['gitlab']['mattermost']['email_apple_push_cert_public'] = nil
+default['gitlab']['mattermost']['email_apple_push_cert_private'] = nil
+
+default['gitlab']['mattermost']['ratelimit_use_rate_limiter'] = true
+default['gitlab']['mattermost']['ratelimit_per_sec'] = 10
+default['gitlab']['mattermost']['ratelimit_memory_store_size'] = 10000
+default['gitlab']['mattermost']['ratelimit_vary_by_remote_addr'] = true
+default['gitlab']['mattermost']['ratelimit_vary_by_header'] = nil
+
+default['gitlab']['mattermost']['privacy_show_email_address'] = true
+default['gitlab']['mattermost']['privacy_show_phone_number'] = true
+default['gitlab']['mattermost']['privacy_show_skype_id'] = true
+default['gitlab']['mattermost']['privacy_show_full_name'] = true
+
+default['gitlab']['mattermost']['team_max_users_per_team'] = 150
+default['gitlab']['mattermost']['team_allow_public_link'] = true
+default['gitlab']['mattermost']['team_allow_valet_default'] = false
+default['gitlab']['mattermost']['team_terms_link'] = '/static/help/configure_links.html'
+default['gitlab']['mattermost']['team_privacy_link'] = '/static/help/configure_links.html'
+default['gitlab']['mattermost']['team_about_link'] = '/static/help/configure_links.html'
+default['gitlab']['mattermost']['team_help_link'] = '/static/help/configure_links.html'
+default['gitlab']['mattermost']['team_report_problem_link'] = '/static/help/configure_links.html'
+default['gitlab']['mattermost']['team_tour_link'] = '/static/help/configure_links.html'
+default['gitlab']['mattermost']['team_default_color'] = '#2389D7'
+
+####
+# Mattermost NGINX
+####
+default['gitlab']['mattermost-nginx'] = default['gitlab']['nginx'].dup
+default['gitlab']['mattermost-nginx']['enable'] = false